round1b/requirements.txt
Flask
openai
pdfplumber
PyMuPDF
python-dotenv


